int to long

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        int[] a = {1, 3, 5, 7, 21};
+        long[] a = {1, 3, 5, 7, 21};
         SalesManager manager = new SalesManager(a);
         System.out.println(manager.max());
     }
